VV3 complete, added formatted output

+++ LU_decom.h
@@ -16,7 +16,7 @@ class LU_decom{
 		int mat[10][10]; //row col
 		double L[10][10]; //row row
 		double U[10][10]; //row col
-		int P[10][10]; // row row
+		double P[10][10]; // row row
 		bool FLAG = false; //check if P_LU has been executed
 };
